requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
djangorestframework==3.7.7
idna==2.7
pytz==2018.5
requests==2.19.1
urllib3==1.23
